Cleanup on the ProTx subdir for the project.Code was getting to long in the tooth in the util.js file and becoming increasingly over-lapping in areas of concern. This refactor does the following:
- Moves all shared utils or other files under a new common folder.
- Moves all data being ingested from files into a new data folder. Later this can become the location of the sqlite3.db flatfile we track.
- Breaks the monolithic utility.js file into three separate files with specific areas of conern:
-- dataUtils.js: any method that munges data to produce new data for use by other methods independent of component.
-- mapUtils.js: any method that is explicit to the map component.
-- IntervalColorScale is now part of this module.
-- plotUtils.js: any method that is explicit to rendering the plot components. These may be further exploded into individual plot type utilities as work progresses on enhancing the plots.
- Deletes the unused files (utils.js, utils.test.js, IntervalColorScale.js).
- Moves the color scale constants all into one common colors.js module under shared.

- Both the mapUtils.js and plotTuils.js files import any dataUtils.js methods necessary.
- All concerns are separated now and grow more independently.

Note: This is in preparation for a major plot component refactor that will be necessary to maintain sanity while iterating rapidly with the cooks teams. A more robust and extensible solution is required in order to keep the work load manageable on the graphing side (at a minimum).
