+++ client/src/components/ProTx/components/common/Variables.css
