+++ client/src/components/ProTx/components/dashboard/MaltreatmentSelector.js
@@ -1,7 +1,7 @@
 import React, { useEffect, useState } from 'react';
 import MultiSelect from 'react-multi-select-component';
 import PropTypes from 'prop-types';
-import { MALTREATMENT } from '../meta';
+import { MALTREATMENT } from '../data/meta';
 import './MaltreatmentSelector.css';
 
 const MaltreatmentSelector = ({ selectedTypes, setSelectedTypes }) => {
